prep for vpn test

+++ scripts/init-firehawk.sh
@@ -65,10 +65,14 @@ else
   echo "...Provision Local VM's"
   $TF_VAR_firehawk_path/scripts/init-openfirehawkserver-020-init.sh $ARGS; exit_test
   echo "...Start Terraform"
-  terraform init -lock=false; exit_test
-  terraform destroy --auto-approve -lock=false; exit_test
+  if [[ "$TF_VAR_init" == true ]]; then
+    terraform init -lock=false; exit_test
+    terraform destroy --auto-approve -lock=false; exit_test
+  fi
   terraform apply --auto-approve -lock=false; exit_test
-  terraform destroy --auto-approve -lock=false; exit_test
+  if [[ "$TF_VAR_destroy_after_deploy" == true ]]; then
+    terraform destroy --auto-approve -lock=false; exit_test
+  fi
   # After this point provisioning will now execute from TF.
   # $TF_VAR_firehawk_path/scripts/init-openfirehawkserver-030-tf-s3user-deadlinercs.sh $ARGS; exit_test
   # $TF_VAR_firehawk_path/scripts/init-openfirehawkserver-040-ssh-routes-nfs-houdini-license-repository.sh $ARGS; exit_test
